# cicd/apispec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "lms-user-service-api",
    "version": "2024-09-04T11:19:56Z"
  },
  "paths": {
...
    "/api/v1/message": {
      "post": {
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empty"
                }
              }
            }
          }
        },
        "security": [
          {
            "lambda-authorizer": []
          }
        ],
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "${service_lambda_function}",
          "responses": {
            "default": {
              "statusCode": "200"
            }
          },
          "passthroughBehavior": "${passthroughbehaviour}",
          "timeoutInMillis": 29000,
          "contentHandling": "CONVERT_TO_TEXT",
          "type": "aws_proxy"
        }
      },
      "get": {
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empty"
                }
              }
            }
          }
        },
        "security": [
          {
            "lambda-authorizer": []
          }
        ],
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "${service_lambda_function}",
          "responses": {
            "default": {
              "statusCode": "200"
            }
          },
          "passthroughBehavior": "${passthroughbehaviour}",
          "timeoutInMillis": 29000,
          "contentHandling": "CONVERT_TO_TEXT",
          "type": "aws_proxy"
        }
      },
      "options": {
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Credentials": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {}
          }
        },
        "x-amazon-apigateway-integration": {
          "type": "mock",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Credentials": "'true'",
                "method.response.header.Access-Control-Allow-Methods": "'OPTIONS,GET,POST,PUT'",
                "method.response.header.Access-Control-Allow-Headers": "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token,X-Portal-Id,Cookie,X-Tenant-ID'",
                "method.response.header.Access-Control-Allow-Origin": "${allow-origins}"
              }
            }
          },
          "requestTemplates": {
            "application/json": "{\"statusCode\": 200}"
          },
          "passthroughBehavior": "${passthroughbehaviour}"
        }
      }
    },
...
    "/api/v1/lmsusers/userByEmail/{email}": {
      "get": {
        "responses": {
          "200": {
            "description": "200 response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empty"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "mls-lambda-authorizer": []
          }
        ],
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "${service_lambda_function}",
          "responses": {
            "default": {
              "statusCode": "200"
            }
          },
          "passthroughBehavior": "${passthroughbehaviour}",
          "timeoutInMillis": 29000,
          "contentHandling": "CONVERT_TO_TEXT",
          "type": "aws_proxy"
        }
      },
      "options": {
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Credentials": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {}
          }
        },
        "x-amazon-apigateway-integration": {
          "type": "mock",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Credentials": "'true'",
                "method.response.header.Access-Control-Allow-Methods": "'OPTIONS,GET'",
                "method.response.header.Access-Control-Allow-Headers": "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token,X-Portal-Id,Cookie,X-Tenant-ID'",
                "method.response.header.Access-Control-Allow-Origin": "${allow-origins}"
              }
            }
          },
          "requestTemplates": {
            "application/json": "{\"statusCode\": 200}"
          },
          "passthroughBehavior": "${passthroughbehaviour}"
        }
      }
    },
...
    "/api/v1/lmsusers/tenantUsersCount": {
      "post": {
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Empty"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response with tenant user counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tenantCodes": {
                      "type": "array",
                      "items": { "type": "string" }
                    },
                    "counts": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "lambda-authorizer": []
          }
        ],
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "${service_lambda_function}",
          "responses": {
            "default": {
              "statusCode": "200"
            }
          },
          "passthroughBehavior": "${passthroughbehaviour}",
          "timeoutInMillis": 29000,
          "contentHandling": "CONVERT_TO_TEXT",
          "type": "aws_proxy"
        }
      },
      "options": {
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": { "type": "string" }
              },
              "Access-Control-Allow-Methods": {
                "schema": { "type": "string" }
              },
              "Access-Control-Allow-Credentials": {
                "schema": { "type": "string" }
              },
              "Access-Control-Allow-Headers": {
                "schema": { "type": "string" }
              }
            },
            "content": {}
          }
        },
        "x-amazon-apigateway-integration": {
          "type": "mock",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Credentials": "'true'",
                "method.response.header.Access-Control-Allow-Methods": "'OPTIONS,PUT,POST,GET'",
                "method.response.header.Access-Control-Allow-Headers": "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token,X-Portal-Id,Cookie,X-Tenant-ID'",
                "method.response.header.Access-Control-Allow-Origin": "${allow-origins}"
              }
            }
          },
          "requestTemplates": {
            "application/json": "{\"statusCode\": 200}"
          },
          "passthroughBehavior": "${passthroughbehaviour}"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Empty": {
        "title": "Empty Schema",
        "type": "object"
      }
    },
    "securitySchemes": {
      "lambda-authorizer": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header",
        "x-amazon-apigateway-authorizer": {
          "type": "request",
          "authorizerUri": "${authorizer_lambda_function}",
          "identitySource": "method.request.header.authorization",
          "authorizerResultTtlInSeconds": 300
        },
        "x-amazon-apigateway-authtype": "Custom"
      },
      "mls-lambda-authorizer": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header",
        "x-amazon-apigateway-authorizer": {
          "type": "request",
          "authorizerUri": "${mls_authorizer_lambda_function}",
          "identitySource": "method.request.header.authorization",
          "authorizerResultTtlInSeconds": 300
        },
        "x-amazon-apigateway-authtype": "Custom"
      },
      "api_key": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header"
      }
    }
  },
...
}